==> Preparing_data/1_Data_cleaning.ipynb <==
# Обработка собранных данных

## Импорт библиотек

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import re

## Открытие данных

df = pd.read_csv('../data.csv', ';')

df.sample(3)

df.info()

## Удаление записей с пропусками важных признаков

important_features = [
    "physical address", "number of rooms", "area of apartment", "number of floors",
    "apartment floor", "price", "repair", "bathroom", "terrace", "year of construction",
    "type of house", 
]

new_df = df.copy()
for feature in important_features:
    new_df = new_df[new_df[feature].isna() == False]

new_df.info()

## Выделение новых признаков из физического адреса

### Получение района города, в котором находится квартира

def get_district(address):
    result = re.search('р-н [а-яА-Я]+', address)
    if result is None:
        return None
    else:
        return result[0][4:]

new_df['district'] = np.array([get_district(item) for item in  new_df['physical address'].values])

print("Найденные районы: " + str(new_df['district'].unique()))

print("Примеры нераспознанных адресов: \n" + '\n'.join(new_df[new_df['district'].isna() == True]['physical address'][:10].values))

Пермский район нас не интересует, поэтому удаляем все эти записи из таблицы

new_df = new_df[new_df['district'].isna() == False]

new_df.info()

### Получение полного адреса без района

def get_full_address(text):
    address = text.split('|')[0]
    replaces = {
        'ш.': 'шоссе',
        'пр-т': 'проспект',
        'б-р': 'бульвар',
        'пр.': 'проезд'
    }
    for key in replaces:
        if key in address:
            address = address.replace(key, replaces[key])
    return address

new_df['full address'] = np.array([get_full_address(item) for item in new_df['physical address'].values])

print(new_df['full address'].unique())

new_df = new_df[new_df['full address'].isna() == False]

new_df.info()

В дальнейшем полный адрес будет использоваться для генерации признаков, связанных с географическим положением квартиры

## Корректировка имеющихся признаков

## Количество комнат

print("Уникальные значения числа комнат: " + ', '.join(new_df['number of rooms'].unique()))

def get_number(string):
    if string == 'студия' or string == 'своб. планировка':
        return 1
    else:
        return int(string)

new_df['number of rooms'] = np.array([get_number(item) for item in new_df['number of rooms'].values])

print("Уникальные значения числа комнат после корректировки: " + str(new_df['number of rooms'].unique()))

### Площадь квартиры

print("Типичные значения для площади: " + ', '.join(new_df['area of apartment'][:10]))

new_df['area of apartment'] = np.array([float(item[:-3]) for item in new_df['area of apartment'].values])

print("Значения для площади после корректировки: " + ', '.join(list(map(str, new_df['area of apartment'].values[:10]))))

### Количество этажей в доме

print("Уникальные значения для числа этажей в доме: \n" + ', '.join(new_df['number of floors'].unique()))

new_df['number of floors'] = np.array([int(re.findall('[\d]+', x)[0]) for x in new_df['number of floors'].values])

print("Уникальные значения для числа этажей в доме после корректировки: \n" + ', '.join(list(map(str, new_df['number of floors'].unique()))))

### Этаж квартиры

new_df['apartment floor'].unique()

new_df['apartment floor'] = new_df['apartment floor'].astype(int)

new_df['apartment floor'].unique()

### Цена

new_df['price'].dtype

Производить манипуляции над ценой не нужно

### Ремонт

print("Уникальные значения для ремонта: " + ', '.join(new_df['repair'].unique()))

Предобработка не требуется, в дальнейшем данный признак будет закодирован методом one hot encoding

### Санузел

print("Уникальные значения для санузла: " + ', '.join(new_df['bathroom'].unique()))

Предобработка не требуется, в дальнейшем данный признак будет закодирован методом one hot encoding

### Вид из окна

print("Уникальные значения для вида из окна: " + str(new_df['view from the windows'].unique()))

Признак малоинформативен, его нужно будет удалить

### Вид балкона

print("Уникальные значения для санузла: " + ', '.join(new_df['terrace'].unique()))

Предобработка не требуется, в дальнейшем данный признак будет закодирован методом one hot encoding

### Год постройки

print("Уникальные значения для года постройки: " + ', '.join(new_df['year of construction'].unique()))

def year(y):
    y = int(y)
    if y < 100:
        return 1900 + y
    elif y < 1000:
        return y*10
    else:
        return y

new_df['year of construction'] = np.array([year(item) for item in new_df['year of construction'].values])

print("Уникальные значения для года постройки после коррекции: " + str(new_df['year of construction'].unique()))

### Количество лифтов

print("Уникальные значения для числа лифтов: " + str(new_df['elevator'].unique()))

print("Количество записей с неизвестным числом лифтов: " + str(len(new_df[new_df['elevator'].isna() == True])))

Для заполнения пропусков воспользуемся будем использовать следующее правило:
1. До 5 этажей в ключительно - нет лифтов
2. От 6 до 9 этажей включительно - 1 лифт
3. От 10 до 19 этажей включительно - 2 лифта
4. От 20 этажей - 3 лифта

def get_n_elevator(floor):
    if floor <= 5:
        return 0
    elif 6 <= floor <= 9:
        return 1
    elif 10 <= floor <= 19:
        return 2
    else:
        return 3

new_df.loc[new_df.elevator.isna(), 'elevator'] = np.array([get_n_elevator(item) for item in new_df[new_df.elevator.isna()]['number of floors'].values])

print("Количество записей с неизвестным числом лифтов: " + str(len(new_df[new_df['elevator'].isna() == True])))

print("Уникальные значения для числа лифтов после коррекции: " + str(new_df['elevator'].unique()))

### Дополнительные услуги

print("Уникальные значения для дополнительных услуг: " + str(new_df['extra'].unique()))

Из этих признаков нас интересует только мусоропровод, консьерж. Если информации нет, то и признаки эти будут отсутствовать у дома.

def get_garbage(text):
    if 'мусоропровод' in text:
        return 1
    else:
        return 0

def get_concierge(text):
    if 'консьерж' in text:
        return 1
    else:
        return 0

new_df['concierge'] = 0
new_df['garbage chute'] = 0

new_df.loc[new_df.extra.notnull(), 'concierge'] = np.array([get_concierge(item) for item in new_df[new_df.extra.notnull()]['extra'].values])
new_df.loc[new_df.extra.notnull(), 'garbage chute'] = np.array([get_garbage(item) for item in new_df[new_df.extra.notnull()]['extra'].values])

### Тип дома

print("Уникальные значения для типа дома: \n" + '\n'.join(new_df['type of house'].unique()))

Замечен тип дома, принимающий экстремальное значение, это нужно исправить

def get_house_type(hous_type):
    if len(hous_type) < 20:
        return hous_type
    else:
        return "кирпичный"

new_df['type of house'] = np.array([get_house_type(item) for item in new_df['type of house'].values])

print("Уникальные значения для типа дома после коррекции: \n" + '\n'.join(new_df['type of house'].unique()))

### Парковка

print("Уникальные значения для парковки: \n" + str(new_df['parking'].unique()))

print("Количество записей с неизвестным типом парковки: " + str(new_df['parking'].isna().sum()))

new_df.loc[new_df.parking.isna(), 'parking'] = ""

Этот признак, несмотря на большое количество пропусков, можно оставить, так как этот признак будет разбит на несколько категорий, исходя из типа парковки, методом one hot encoding.

## Удаление ненужных столбцов

new_df.columns

column_for_delete = ["physical address", "link", "view from the windows", "extra"]

for column in column_for_delete:
    new_df = new_df.drop(column, axis=1)

new_df.columns

new_df.index = np.arange(len(new_df))

## Итог первичной обработки

new_df.info()

new_df.to_csv('../clean_data.csv', sep=';', index=False)

__В ходе первой обработки были:__
1. удалены записи с пропусками важных признаков
2. выделены признаки района и полного адреса без района
3. рассмотрены и скорректированы некоторые признаки
4. удалены ненужные столбцы
