==> Preparing_data/3_Categorical_features.ipynb <==
# Преобразование категориальных признаков очищенных данных

## Импорт библиотек

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import sklearn
import re

## Открытие данных

df = pd.read_csv('../clean_data_without_outliers.csv', sep=';')

df.sample(3)

df.info()

## Обнаружение категориальных признаков

cat_features = []
for column in df.columns:
    if str(df[column].dtype) == "object":
        cat_features.append(column)
print("Количество признаков: " + str(len(cat_features)))

cat_features

Все признаки кроме полного адреса мы закодируем с помощью метода one hot encoding

### Обработка ремонта, типа ванной, типа балкона, типа стен дома, района

simple_cat_features = ['repair', 'bathroom', 'terrace', 'type of house', 'district']

for cat_feature in simple_cat_features:
    print(cat_feature + str(df[cat_feature].unique()))

Данные признаки содержат только одно значение, поэтому их можно обрабатывать единым образом

from sklearn.preprocessing import OneHotEncoder

def get_one_hot_df(df, column):
    encoder = OneHotEncoder()
    new_df = pd.DataFrame(encoder.fit_transform(df[[column]]).toarray())
    new_df.columns = [column + '_' + feature[3:] for feature in encoder.get_feature_names()]
    return new_df

for cat_feature in simple_cat_features:
    one_hot_df = get_one_hot_df(df, cat_feature)
    df = pd.merge(df, one_hot_df, left_index=True, right_index=True)
    df = df.drop(cat_feature, axis=1)

df.info()

### Кодирование типа парковки

df.loc[df.parking.isna(), 'parking'] = ""

unique_parking = set()
for parking_set in df.parking.unique():
    for parking_type in parking_set.split(','):
        unique_parking.add(parking_type.strip())
unique_parking.remove('')

unique_parking

def get_parking_type(string, parking_type):
    return float(parking_type in string)

for parking_type in unique_parking:
    df["parking_" + parking_type] = get_parking_type(df.parking, parking_type)
df = df.drop('parking', axis=1)

## Итоги преобразования категориальных признаков

df.info()

df.to_csv('../clean_categorizated_data_without_outliers.csv', sep=';', index=False)

__В ходе третьей обработки были:__
1. трансформированы категориальные признаки по методу one hot encoding
