==> Preparing_data/4_Coordinates_generation.ipynb <==
# Генерация координат квартир

## Импорт библиотек

import requests

import json
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import sklearn
import re

from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
import time

## Открытие данных

df = pd.read_csv('../clean_categorizated_data_without_outliers.csv', sep=';')

df.sample(3)

df.info()

## Получение координат дома квартиры

geo_coder_url = "https://geotree.ru/geocoder"

driver = webdriver.Chrome(ChromeDriverManager().install())

coords = np.zeros(shape=(len(df), 2))
no_coords_dict = dict()

for idx, address in enumerate(df['full address']):
    if idx%100 == 0:
        print(f"{idx}/{len(df)}")
    try:
        # Обращаемся к сайту-геокедру
        driver.get(geo_coder_url)
        # Находим элемент для вставки адреса и помещаем туда адрес
        input_address = driver.find_element_by_id("input_address")
        input_address.clear()
        input_address.send_keys(address)
        time.sleep(1)
        # Извлекаем результат и преобразуем в словарь
        result = driver.find_element_by_id("div_result_json").find_element_by_tag_name('pre')
        text = result.text.replace('\n', '')
        info = json.loads(text)
        # Достаем из словаря координаты объекта
        lat = info[0].get('geo_center').get('lat')
        lon = info[0].get('geo_center').get('lon')
        # Сохраняем координаты
        coords[idx][0] = lat
        coords[idx][1] = lon
    except Exception as ex:
        print("Обнаружена ошибка с адресом номер " + str(idx) + " " + address)
        print(ex)
        if address in no_coords_dict:
            no_coords_dict[address].append(idx)
        else:
            no_coords_dict[address] = [idx]
        print(no_coords_dict[address])
        print("--------------------")
        continue

driver.close()

## Вручную дополняем самые распространенные адреса

driver = webdriver.Chrome(ChromeDriverManager().install())

for address in no_coords_dict:
    if not no_coords_dict[address]:
        continue
    elif len(no_coords_dict[address]) >= 2:
        try:
            # Обращаемся к сайту-геокедру
            driver.get(geo_coder_url)
            # Находим элемент для вставки адреса и помещаем туда адрес
            input_address = driver.find_element_by_id("input_address")
            input_address.clear()
            input_address.send_keys(address)
            time.sleep(2)
            # Извлекаем результат и преобразуем в словарь
            result = driver.find_element_by_id("div_result_json").find_element_by_tag_name('pre')
            text = result.text.replace('\n', '')
            info = json.loads(text)
            # Достаем из словаря координаты объекта
            lat = info[0].get('geo_center').get('lat')
            lon = info[0].get('geo_center').get('lon')
            # Сохраняем координаты
            for idx in no_coords_dict[address]:
                coords[idx][0] = lat
                coords[idx][1] = lon
            no_coords_dict[address] = None
        except Exception as ex:
            print("Обнаружена ошибка с адресом номер " + address)
            print(ex)
            string = input().split(', ')
            lat = float(string[0].strip())
            lon = float(string[1].strip())
            if lon == 0 and lat == 0:
                break
            for idx in no_coords_dict[address]:
                coords[idx][0] = lat
                coords[idx][1] = lon
            no_coords_dict[address] = None
            print("--------------------")
            continue

driver.close()

## Сохраняем координаты в таблицу

lat = coords[:, 0].copy()
lon = coords[:, 1].copy()

lat.shape, lon.shape

df['lat'] = lat
df['lon'] = lon

df = df.loc[df.lat != 0]

## Итоги генерации координат

df.info()

df.to_csv('../clean_categorizated_data_without_outliers_with_coordinates.csv', sep=';', index=False)

__В ходе четвертой обработки были:__
1. получены значения координат квартир
