==> Preparing_data/5_Amenity_feature_generation.ipynb <==
# Получение данных об инфраструктуре поблизости от квартиры

## Импорт библиотек

import requests
import json
import time
import pandas as pd
import numpy as np

## Открытие данных

df = pd.read_csv('../clean_categorizated_data_without_outliers_with_coordinates.csv', sep=';')

df.sample(3)

df.info()

## Парсинг данных координат для определенных категорий

Запускаем алгоритм парсинга несколько раз, чтобы спарсить все подкатегории

all_categories = {
    "edu":{
        "amenities": ["college", "kindergarten", "library", "school", "university"],
        "data": set()
    },
    "health":{
        "amenities": ["clinic", "hospital"],
        "data": set()
    },
    "culture":{
        "amenities": ["cinema", "fountain", "theatre"],
        "data": set()
    },
    "eat":{
        "amenities": ["cafe", "fast_food", "pub", "restaurant"],
        "data": set()
    },
}

overpass_url = "http://overpass-api.de/api/interpreter"

for category in all_categories:
    if type(all_categories[category]) is set:
        continue
    for amenity in all_categories[category]['amenities']:
        time.sleep(3)
        overpass_query = f"""
        [out:json];
        node(57.5,55,58.3,56.6)[amenity={amenity}];out;
        """
        response = requests.get(overpass_url, 
                                params={'data': overpass_query})
        try:
            data = response.json()
            for element in data['elements']:
              if element['type'] == 'node':
                lon = element['lon']
                lat = element['lat']
                all_categories[category]['data'].add((lat, lon))
            all_categories[category]['amenities'].remove(amenity)
            if not all_categories[category]['amenities']:
                del all_categories[category]['amenities']
                all_categories[category] = all_categories[category]['data']
        except Exception as ex:
            print(amenity)
            print(ex)

for key in all_categories:
    print(key + ": " + str(len(all_categories[key])))

for key in all_categories:
    all_categories[key] = list(all_categories[key])

with open("../amenities.json", 'w') as f:
    json.dump(all_categories, f)

with open("../amenities.json", 'r') as f:
    data = json.load(f)

for key in data:
    print(key + ": " + str(len(data[key])))

## Расстояние между объектами через координатами

from math import sin, cos, asin, sqrt, pi

def get_dist(llong1, llat1, llong2, llat2):
    
    rad = 6372795
    
    #в радианах
    lat1 = llat1*pi/180.
    lat2 = llat2*pi/180.
    long1 = llong1*pi/180.
    long2 = llong2*pi/180.
    
    #косинусы и синусы широт и разницы долгот
    delta_long = long2 - long1
    delta_lat = lat2 - lat1
    
    #вычисления длины большого круга
    ad = 2 * asin(sqrt(sin(delta_lat/2)**2 + cos(lat1)*cos(lat2)*sin(delta_long/2)**2))
    dist = ad*rad
    
    return dist

## Генерация новых признаков - услуг вблизи квартиры

Для каждой категории вычисляем расстояние от квартиры до объекта, где предоставляют услугу данной категории

new_df = df.copy()
categories_array = np.zeros(shape=(len(new_df), 4))

for i in range(len(new_df)):
    appartment = new_df.loc[i]
    appartment_lat = appartment['lat']
    appartment_lon = appartment['lon']
    for cat_id, category in enumerate(data):
        for coordinates in data[category]:
            object_lat = coordinates[0]
            object_lon = coordinates[1]
            dist = get_dist(
                llong1=appartment_lon,
                llat1=appartment_lat,
                llong2=object_lon,
                llat2=object_lat
            )
            if dist < 5000:
                categories_array[i, cat_id] +=1

for cat_id, category in enumerate(data):
    new_df[category] = categories_array[:, cat_id]

new_df.sample(5)

## Генерация новых признаков - расстояние до центра

perm_esplanade_lat = 58.010455
perm_esplanade_lon = 56.229443

new_df['distance'] = 0

for i in range(len(new_df)):
    appartment = new_df.loc[i]
    appartment_lat = appartment['lat']
    appartment_lon = appartment['lon']
    dist = get_dist(
        llong1=appartment_lon,
        llat1=appartment_lat,
        llong2=perm_esplanade_lon,
        llat2=perm_esplanade_lat
    )
    new_df.loc[i, 'distance'] = round(dist, 2)

## Итоги генерации признаков

new_df.info()

new_df.to_csv('../clean_categorizated_data_without_outliers_with_coordinates_and_amenities.csv', sep=';', index=False)

__В ходе пятой обработки были:__
1. добавлены расстояния до центра
2. добавлены признаки наличия услуг в радиусе около 5 км
