==> Preparing_data/6_For_NN.ipynb <==
# Подготовка данных к обучению на нейронной сети

## Импорт библиотек

import pandas as pd
import numpy as np

## Открытие данных

df = pd.read_csv('../clean_categorizated_data_without_outliers_with_coordinates_and_amenities.csv', sep=';')

df.sample(3)

df.info()

## Удаление из данных полного адреса

df = df.drop('full address', axis=1)

## Приведение типов к float32

df = df.astype(np.float32)

## Удаление признака парковки

df = df.drop(['parking_за шлагбаумом во дворе', 'parking_открытая во дворе', 
              'parking_наземная многоуровневая', 'parking_подземная'], axis=1)

## Итоги подготовки данных

df.info()

df.sample(3)

df.to_csv('../final_data.csv', sep=';', index=False)

__В ходе шестой обработки были:__
1. отброшены не числовый признаки
2. все числовые признаки были приведены к типу float
После испытаний были:
1. удалены признаки парковки
